don't use a custom dialer
Since Kubernetes 1.20, HTTP2 connections run a periodic ping every
30 seconds to detect stale connections that detect if the connection
was broken.

This has the advantage over kernel parameters that the connection is
always alive and you can detect dead peers. Using the Kernel parameters
is very expensive and agressive, since once the connection is established,
it stops sending TCP keepalives, and when the TCP_USER_TIMEOUT kicks off for
not having activity on the connection, it tears down the connection
ungracefully, without knowing if the peer was really dead, with a high
risk of closing legit connections.

Golang default parameters for dialer are good, also, overriding the Dial
function in client-go automatically disables transport caching.

Signed-off-by: Antonio Ojea <aojea@redhat.com>